+++ src/app/main/content/apps/master-data/range-price/range-price-list/range-price-list.component.ts
@@ -6,6 +6,8 @@ import { FormGroup, FormBuilder } from '@angular/forms';
 import * as FileSaver from 'file-saver';
 import { UserService } from '@fuse/directives/users/users.service';
 import { Functions } from '@fuse/core/function';
+import { MatDialogRef, MatDialog } from '@angular/material';
+import { FuseConfirmDialogComponent } from '@fuse/components/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'range-price-list',
@@ -30,12 +32,14 @@ export class RangePriceListComponent implements OnInit {
   hasDeleteUserPermission = false;
   private hasViewUserPermission = false;
   service;
+  dialogRef: MatDialogRef<FuseConfirmDialogComponent>;
 
   constructor(
     private router: Router,
     private rangePriceListService: RangePriceListService,
     private formBuilder: FormBuilder,
     private toastyService: ToastyService,
+    public dialog: MatDialog,
     private _user: UserService,
     private _Func: Functions,
     private toastyConfig: ToastyConfig
@@ -131,15 +135,23 @@ export class RangePriceListComponent implements OnInit {
     } else if (this.selected.length > 1) {
       this.toastyService.error('Please select one item.');
     } else {
-      this.rangePriceListService.deleteRangePrice(this.selected[0]['id_temp']).subscribe((data) => {
-        this.toastyService.success(data['message']);
-        setTimeout(
-          () => {
-            this.getList();
-            this.selected = [];
-          },
-          700
-        );
+      this.dialogRef = this.dialog.open(FuseConfirmDialogComponent);
+      this.dialogRef.componentInstance.confirmMessage = 'Are you sure you want to delete?';
+      this.dialogRef.afterClosed().subscribe(result => {
+        if (result) {
+          this.rangePriceListService.deleteRangePrice(this.selected[0]['id_temp']).subscribe((data) => {
+            this.toastyService.success(data['message']);
+            setTimeout(
+              () => {
+                this.getList();
+                this.selected = [];
+              },
+              700
+            );
+          });
+        } else {
+        }
+        this.dialogRef = null;
       });
     }
   }
